fix: Remove ESP32-H4 support
Since ESP32-H2 has been released, support for the engineering variant
of the chip is no longer needed.

The spot in the enum has been reserved for future use in order to avoid
a possible breaking change for users who used the exact numeric values
of the target_chip_t enum.