feat: Add the ability for ESP ports to not initialize peripherals
This adds the ability of the ESP serial and SPI ports to not initialize
the peripheral. This enables using the peripherals for multiple devices
or multiple uses for one device (e.g. flashing and monitoring).

The boolean flag added to config structures will be false by default
so no changes are required for existing code.