change: Refactor the project CMakeLists
This is done to reduce redundancy for interface and option handling between different
ports and brings them to the start of the CMakeLists, as it
is more relevant to users than port handling most of the time.

A macro is introduced to streamline option creation and reduce the chance of error.