test: Drop host tests for now
This drops host tests due to the flawed mocking approach resulting in
an inability to produce timeouts, as well as lack of comprehensiveness
of the test suite and qemu and (soon) target tests covering all
tested usecases.

This isn't a ban of mocking for host testing, we should revisit it
in the future for cases where it might not be easy to reproduce certain
behavior with target or qemu tests and for unit testing of parts of
the library.