fix: Update USB reset sequence
This commit fixes the USB reset sequence for the ESP32 USB CDC ACM port.
loader_port_reset_target() function kept the target in reset state so it
was not be possible to use it standalone. This commit also makes the
entering the bootloader to go through state where BOOT and RESET are
both set.