fix: INVALID_ARG error when ESP32P4 used as a host
This fixes the issue where the ESP32P4 would return an INVALID_ARG error
when acting as a host. This was due to new cache of ESP32P4 that
requires aligning to 64 bytes. This commit adds the flag to sdmm_host
config that creates a 64 byte aligned buffer for the host when ESP32P4
is used.