ci: Add codespell and markdown linter pre-commit hooks and update existing ones
This also removes pre-commit github workflow as it will be enabled by github
autoupdate feature.

This commit adds pre-commit hooks for codespell and markdown linting to
the project. It also updates the existing pre-commit configuration and
runs the hooks on the current codebase to ensure consistency and correctness.
This also adds section that enables github to autoupdate the pre-commit hooks.

Closes https://github.com/espressif/esp-serial-flasher/issues/145